refactor: hooks directory architecture change
