+++ src/pages/point-shop/products/[pointProductId]/detail.tsx
@@ -2,7 +2,7 @@ import DeliveryAddress from '@/components/shop-screen/delivery-address'
 import PointDescription from '@/components/shop-screen/point-description'
 import ExchangeProduct from '@/components/shop-screen/exchange-product'
 import PageLayOut from '@/components/common/page-layout'
-import useProduct from '@/hooks/use-product'
+import useProduct from '@/hooks/product/use-product'
 import { useUserStatus } from '@/hooks/use-user-status'
 import { useState } from 'react'
 import { useNavigate, useParams } from 'react-router-dom'
